feat(tool): Add template support and case-insensitivity to generator
Makes the parsing of `#+BEGIN/...` and `#+END/...` blocks case-insensitive.

Introduces a `--template` command-line option to `generate_from_org.dart`. When a template class name is provided, the generated slides will use that template to wrap their content. If no template is specified, the script defaults to the standard `FlutterDeckSlide` classes.

The `build` method generation logic has been updated to ensure it correctly returns a `FlutterDeckSlide` and calls the template's `build` method when a template is used.

Fixes a bug that added a redundant `const` keyword to the `FlutterDeckSlideConfiguration` in the generated code.